requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
DateTime==4.3
Flask==1.1.1
gcloud==0.18.3
googleapis-common-protos==1.6.0
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==4.1.2
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycryptodome==3.9.0
Pyrebase4==4.3.0
python-jwt==2.0.1
pytz==2019.2
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
zope.interface==4.6.0